requirements.txt
python-dotenv
requests
webdriver-manager
selenium
google-generativeai
SQLAlchemy
Django
djangorestframework
sqlalchemy_paginator
oauthlib
pusher
regex
python-gitlab